MVP for creating a hosted zone record and externalname svc

+++ controllers/vpcendpoint/cleanup.go
@@ -18,6 +18,10 @@ package vpcendpoint
 
 import (
 	"context"
+	"fmt"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	route53Types "github.com/aws/aws-sdk-go-v2/service/route53/types"
 	avov1alpha2 "github.com/openshift/aws-vpce-operator/api/v1alpha2"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -25,31 +29,34 @@ import (
 
 // cleanupAwsResources cleans up AWS resources associated with a VPC Endpoint.
 func (r *VpcEndpointReconciler) cleanupAwsResources(ctx context.Context, resource *avov1alpha2.VpcEndpoint) error {
-	if meta.IsStatusConditionTrue(resource.Status.Conditions, avov1alpha2.AWSCustomDnsCondition) {
-		//resourceRecord, err := r.generateRoute53Record(ctx, resource)
-		//if err != nil {
-		//	return err
-		//}
+	if meta.IsStatusConditionTrue(resource.Status.Conditions, avov1alpha2.AWSRoute53RecordCondition) {
+		resourceRecord, err := r.generateRoute53Record(ctx, resource)
+		if err != nil {
+			return err
+		}
 
-		//hostedZone, err := r.awsClient.ListHostedZonesByVPC(ctx, r.clusterInfo.domainName)
-		//if err != nil {
-		//	return err
-		//}
-		//
-		//input := &route53Types.ResourceRecordSet{
-		//	Name:            aws.String(fmt.Sprintf("%s.%s", resource.Spec.CustomDns.Route53PrivateHostedZone.Record.Hostname, *hostedZone.Name)),
-		//	ResourceRecords: []route53Types.ResourceRecord{*resourceRecord},
-		//	TTL:             aws.Int64(300),
-		//	Type:            route53Types.RRTypeCname,
-		//}
+		resp, err := r.awsClient.GetHostedZone(ctx, resource.Status.HostedZoneId)
+		if err != nil {
+			return err
+		}
 
-		//r.log.V(0).Info("Deleting Route53 Hosted Zone Record")
-		//if _, err := r.awsClient.DeleteResourceRecordSet(ctx, input, *hostedZone.Id); err != nil {
-		//	return err
-		//}
+		if resp.HostedZone != nil {
+			r.log.V(0).Info("Deleting Route53 Hosted Zone Record")
+			input := &route53Types.ResourceRecordSet{
+				Name:            aws.String(fmt.Sprintf("%s.%s", resource.Spec.CustomDns.Route53PrivateHostedZone.Record.Hostname, *resp.HostedZone.Name)),
+				ResourceRecords: []route53Types.ResourceRecord{*resourceRecord},
+				TTL:             aws.Int64(300),
+				Type:            route53Types.RRTypeCname,
+			}
 
+			if _, err := r.awsClient.DeleteResourceRecordSet(ctx, input, *resp.HostedZone.Id); err != nil {
+				return err
+			}
+		}
+
+		resource.Status.ResourceRecordSet = ""
 		meta.SetStatusCondition(&resource.Status.Conditions, metav1.Condition{
-			Type:    avov1alpha2.AWSCustomDnsCondition,
+			Type:    avov1alpha2.AWSRoute53RecordCondition,
 			Status:  metav1.ConditionFalse,
 			Reason:  "Deleted",
 			Message: "Deleted Route53 Hosted Zone Record",
@@ -62,10 +69,12 @@ func (r *VpcEndpointReconciler) cleanupAwsResources(ctx context.Context, resourc
 	}
 
 	if resource.Status.HostedZoneId != "" {
-		if _, err := r.awsClient.DeleteHostedZone(ctx, resource.Status.HostedZoneId); err != nil {
-			return err
+		// Only delete a Route53 Private Hosted Zone if AVO created it
+		if resource.Spec.CustomDns.Route53PrivateHostedZone.DomainName != "" {
+			if _, err := r.awsClient.DeleteHostedZone(ctx, resource.Status.HostedZoneId); err != nil {
+				return err
+			}
 		}
-
 		resource.Status.HostedZoneId = ""
 		if err := r.Status().Update(ctx, resource); err != nil {
 			r.log.V(0).Error(err, "failed to update status")
@@ -104,6 +113,7 @@ func (r *VpcEndpointReconciler) cleanupAwsResources(ctx context.Context, resourc
 		}
 	}
 
+	r.log.V(0).Info("AWS cleanup complete")
 	return nil
 }
